# openapi.yaml
openapi: 3.0.3
info:
  title: StrathShop API
  description: |
    API exposed by Group 8 (StrathShop) to Group 9 (the Strathmore Mental
    Health App). This contract covers read access to Listings and Categories,
    plus one scoped write endpoint for updating a listing's availability
    status after a booking action on their side. No student, session, or user
    account data is exposed.

    **Conventions**

    - **Identifiers are lowercase.** Every enum value, and the `category`
      field on a listing, is always returned in lowercase.
    - **Input is case-insensitive.** The `category` query parameter and the
      `status` field in a PATCH body are matched without regard to case and
      with surrounding whitespace trimmed, so `Wellness`, `wellness` and
      ` WELLNESS ` are the same value. Responses are always lowercase.
    - **Capitalising for display is the consumer's job.** Any value you can
      filter or compare on is returned lowercase, `Category.label` included.
      The one exception is `condition`, which is free text shown as stored
      ("Like New") and is never used as an identifier.
    - **Path ids must be positive integers.** A non-integer id is treated as
      not found and returns 404, not 400.
  version: 4.0.0

servers:
  - url: http://localhost:5000/api
    description: Local development server

paths:
  /listings:
    get:
      summary: Return marketplace listings, optionally filtered by category
      description: >
        Returns listings in the wellness or counseling category, used to
        surface self-care items and bookable on-campus services.

        Omitting the category parameter returns the full StrathShop
        catalogue, which includes categories outside this contract's scope
        (Textbooks, Electronics, Fashion, Furniture, Services, Food). Those
        listings share the same shape but their category value will not be
        wellness or counseling. Consumers of this contract should always
        pass a category.
      parameters:
        - name: category
          in: query
          required: false
          description: >
            Which listing category to return. Matched case-insensitively and
            trimmed, so `Wellness` and `wellness` behave identically. Omit to
            return the full catalogue.
          schema:
            type: string
            enum: [wellness, counseling]
            example: wellness
      responses:
        '200':
          description: >
            Listings matching the given category. The schemas below describe
            the filtered responses this contract covers.
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/WellnessListing'
                    - $ref: '#/components/schemas/CounselingListing'
        '400':
          description: category was supplied but is not one of the allowed values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /listings/{id}:
    get:
      summary: Return full details for a single listing
      description: >
        Used to check a listing's current price and availability before
        recommending it. An id belonging to a listing outside this
        contract's categories returns the same shape with a different
        category value.
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The listing's internal ID. Must be a positive integer; any other
            value returns 404.
          schema:
            type: integer
            example: 14
      responses:
        '200':
          description: Listing found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/WellnessListing'
                  - $ref: '#/components/schemas/CounselingListing'
        '404':
          description: No listing found with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /listings/{id}/availability:
    patch:
      summary: Update a listing's availability status
      description: >
        Called when a booking action in the Strathmore Mental Health App
        consumes a listing (e.g. a doctor accepting an appointment tied to a
        counseling listing, or a wellness item being claimed). Scoped to the
        status field only; it cannot change a listing's title, price, or
        description.
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The listing's internal ID. Must be a positive integer; any other
            value returns 404.
          schema:
            type: integer
            example: 21
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Availability'
      responses:
        '200':
          description: Availability status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '400':
          description: status is missing or not one of the allowed values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No listing found with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories:
    get:
      summary: Return the list of available marketplace categories
      responses:
        '200':
          description: All available categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  /categories/{id}:
    get:
      summary: Return a single category's details by ID
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The category's internal ID. Must be a positive integer; any other
            value returns 404.
          schema:
            type: integer
            example: 8
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: No category found with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    WellnessListing:
      type: object
      description: A physical wellness or over-the-counter item for sale.
      properties:
        id:
          type: integer
          example: 14
        title:
          type: string
          example: 'Ibuprofen 200mg (24 tablets)'
        price:
          type: number
          format: double
          description: Price in KES. Always a JSON number, never a string.
          example: 250.00
        description:
          type: string
          nullable: true
          description: May be null; not every listing has a description.
          example: 'Sealed box, unopened. Pickup available on campus.'
        category:
          type: string
          enum: [wellness]
          description: Always lowercase, regardless of how it is displayed elsewhere.
          example: wellness
        condition:
          type: string
          enum: [Like New, Good, Fair, 'N/A']
          example: 'Like New'
        status:
          type: string
          enum: [available, booked]
          description: >
            Current availability. Changed via
            PATCH /listings/{id}/availability.
          example: available
        imageUrl:
          type: string
          nullable: true
          example: 'https://placehold.co/400x300?text=Ibuprofen'
        sellerName:
          type: string
          example: 'Strathmore Campus Pharmacy'
        sellerCampus:
          type: string
          example: 'Strathmore University'
        datePosted:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the listing was created.
          example: '2026-09-11T08:00:00.000Z'
      required:
        - id
        - title
        - price
        - description
        - category
        - condition
        - status
        - imageUrl
        - sellerName
        - sellerCampus
        - datePosted

    CounselingListing:
      type: object
      description: A bookable health or counseling service offered on campus.
      properties:
        id:
          type: integer
          example: 21
        title:
          type: string
          example: '1-on-1 Counseling Session'
        price:
          type: number
          format: double
          description: Price in KES. Always a JSON number, never a string.
          example: 1500.00
        description:
          type: string
          nullable: true
          example: '50-minute session with a licensed campus counselor.'
        category:
          type: string
          enum: [counseling]
          example: counseling
        condition:
          type: string
          enum: ['N/A']
          description: Not meaningful for services; always "N/A".
          example: 'N/A'
        status:
          type: string
          enum: [available, booked]
          description: >
            Current availability. Changed via
            PATCH /listings/{id}/availability.
          example: available
        providerName:
          type: string
          description: The doctor or counselor offering this service.
          example: 'Dr. Wanjiku Mwangi'
        imageUrl:
          type: string
          nullable: true
          example: 'https://placehold.co/400x300?text=Counseling+Session'
        sellerName:
          type: string
          description: Same value as providerName for counseling listings.
          example: 'Dr. Wanjiku Mwangi'
        sellerCampus:
          type: string
          example: 'Strathmore University'
        datePosted:
          type: string
          format: date-time
          example: '2026-09-11T08:00:00.000Z'
      required:
        - id
        - title
        - price
        - description
        - category
        - condition
        - status
        - providerName
        - imageUrl
        - sellerName
        - sellerCampus
        - datePosted

    Availability:
      type: object
      properties:
        status:
          type: string
          enum: [available, booked]
          description: >
            Accepted case-insensitively on a request ("Booked" is valid) and
            always returned in lowercase.
          example: booked
      required: [status]

    Category:
      type: object
      properties:
        id:
          type: integer
          example: 8
        label:
          type: string
          description: >
            Lowercase category name. The same value appears as `category` on a
            listing and is what the ?category= filter expects. Capitalise it
            for display.
          example: wellness
      required: [id, label]

    Error:
      type: object
      properties:
        error:
          type: string
          example: 'Listing not found'
      required: [error]